enhance: refine variable-length-type memory usage(#38736)

+++ internal/core/src/common/Common.h
@@ -51,7 +51,7 @@ SetDefaultExecEvalExprBatchSize(int64_t val);
 struct BufferView {
     struct Element {
         const char* data_;
-        uint64_t* offsets_;
+        uint32_t* offsets_;
         int start_;
         int end_;
     };
